requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
langchain-community
langchain-core
chromadb
unstructured
pdfminer.six
python-docx
